+++ packages/omniston-sdk/src/helpers/wrapErrorsAsync.ts
@@ -1,4 +1,4 @@
-import { wrapError } from "@/helpers/wrapError";
+import { wrapError } from "../helpers/wrapError";
 
 export async function wrapErrorsAsync<T>(fn: () => Promise<T>): Promise<T> {
   try {