+++ packages/omniston-sdk/vitest.config.ts
@@ -1,7 +1,4 @@
 /// <reference types="vitest" />
-import tsconfigPaths from "vite-tsconfig-paths";
 import { defineConfig } from "vitest/config";
 
-export default defineConfig({
-  plugins: [tsconfigPaths()],
-});
+export default defineConfig({});